--- a/packages/theme-tasks/lib/utils/nodesass-packageimporter.js
+++ b/packages/theme-tasks/lib/utils/nodesass-packageimporter.js
@@ -1,9 +1,6 @@
 const path = require("path");
 
-const EMPTY_IMPORT = {
-    file: "",
-    contents: ""
-};
+const EMPTY_IMPORT = {};
 
 const imported = new Set();